pull out the "add to calendar" functionality to all be in the same place
