+++ modules/add-to-device-calendar/index.js
@@ -1,0 +1,3 @@
+// @flow
+
+export {AddToCalendar} from './add-to-calendar'
